+++ CupheadOnline/Plugin.cs
@@ -517,6 +517,6 @@ namespace CupheadOnline
     {
         public const string GUID    = "com.cupheadonline.mod";
         public const string NAME    = "CupHeads";
-        public const string VERSION = "1.2.32";
+        public const string VERSION = "1.2.33";
     }
 }